Feature/create game

See merge request cash-flow-team/cash-flow!37

+++ client/lib/models/domain/account_state.dart
@@ -9,7 +9,7 @@ class AccountState {
         assert(cashFlow != null),
         assert(credit != null);
 
-  final int cash;
-  final int cashFlow;
-  final int credit;
+  final double cash;
+  final double cashFlow;
+  final double credit;
 }
