+++ client/lib/models/network/responses/possessions_state/assets/debenture_asset_response_model.dart
@@ -21,8 +21,8 @@ class DebentureAssetResponseModel {
   @JsonKey(name: 'type', defaultValue: AssetType.other)
   final AssetType type;
   final int count;
-  final int nominal;
-  final int averagePrice;
+  final double nominal;
+  final double averagePrice;
   final double profitabilityPercent;
 
   Map<String, dynamic> toJson() => _$DebentureAssetResponseModelToJson(this);
